Rename optionalProfileField method
