+++ app/model/Name.scala
@@ -1,7 +1,7 @@
 package model
 
 import com.okta.sdk.resource.model.User as OktaUser
-import model.User.optionalProfileField
+import utils.StringHelper.nonNullNonEmpty
 
 case class Name(title: Option[String], firstName: Option[String], lastName: Option[String])
 
@@ -9,8 +9,8 @@ object Name {
 
   def fromOktaUser(oktaUser: OktaUser): Name =
     Name(
-      title = optionalProfileField(oktaUser.getProfile.getTitle),
-      firstName = optionalProfileField(oktaUser.getProfile.getFirstName),
-      lastName = optionalProfileField(oktaUser.getProfile.getLastName)
+      title = nonNullNonEmpty(oktaUser.getProfile.getTitle),
+      firstName = nonNullNonEmpty(oktaUser.getProfile.getFirstName),
+      lastName = nonNullNonEmpty(oktaUser.getProfile.getLastName)
     )
 }
