+++ app/services/OktaUserService.scala
@@ -5,11 +5,11 @@ import com.okta.sdk.resource.client.{ApiClient, ApiException}
 import model.User
 import play.api.Logging
 import play.api.libs.ws.WSClient
-import services.FutureHelper.*
+import utils.FutureHelper.tryAsync
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Failure
-import scala.util.chaining.*
+import scala.util.chaining.scalaUtilChainingOps
 
 class OktaUserService(userApi: UserApi, orgUrl: String, ws: WSClient)(implicit ctx: ExecutionContext) extends Logging {
